Change column type for trips distance to integer

+++ db/migrate/20190915144818_add_distance_to_trips.rb
@@ -1,5 +1,5 @@
 class AddDistanceToTrips < ActiveRecord::Migration[6.0]
   def change
-    add_column :trips, :distance, :float
+    add_column :trips, :distance, :integer
   end
 end
